refactor(dashboards): consolidate widget constants module
Move grid column count and widget limits into products/dashboards/backend/constants.py
